bug fix for 404 routing
New behaviour :
-> if there is no handler to match a request, a 404 header is provided
+ the request handler is fooled into believing that the request was for
the 404 url specified in the configuration. Thus the framework includes
the 404 handler associated to that url, wherever it might be.
On the client site, the URL not found is still the one showed, the
error code is a proper 404 but the content is a 404 handler of our
choice.